requirements.txt
requests~=2.32.3
pytz~=2025.2
dotenv~=0.9.9
python-dotenv~=1.1.0
tabulate~=0.9.0
python-dateutil~=2.9.0.post0